Slightly clearer docs
Thanks so much for this theme and these great docs on it! It’s my first time installing a visual studios theme, and the guide on how to do it saved me some headaches. One headache I still found was that the preferences option was under `Code` not `File` (on a mac), so here's a quick PR to update that. Hope this is helpful.